# Improve journal throughput when journalSyncData is disabled.