--- a/bookkeeper-server/src/main/java/org/apache/bookkeeper/bookie/Journal.java
+++ b/bookkeeper-server/src/main/java/org/apache/bookkeeper/bookie/Journal.java
@@ -45,9 +45,14 @@ import java.util.List;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
+import java.util.concurrent.LinkedBlockingQueue;
+import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
+import org.apache.bookkeeper.bookie.Journal.BufferedChannelBuilder;
+import org.apache.bookkeeper.bookie.Journal.ForceWriteRequest;
+import org.apache.bookkeeper.bookie.Journal.JournalScanner;
+import org.apache.bookkeeper.bookie.Journal.LastLogMark;
 import org.apache.bookkeeper.bookie.LedgerDirsManager.NoWritableLedgerDirException;
 import org.apache.bookkeeper.bookie.stats.JournalStats;
 import org.apache.bookkeeper.common.collections.BlockingMpscQueue;
@@ -61,6 +66,7 @@ import org.apache.bookkeeper.stats.OpStatsLogger;
 import org.apache.bookkeeper.stats.StatsLogger;
 import org.apache.bookkeeper.util.IOUtils;
 import org.apache.bookkeeper.util.MathUtils;
+import org.apache.commons.lang.mutable.MutableLong;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -369,7 +375,8 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
         private long logId;
         private long enqueueTime;
 
-        public int process(boolean shouldForceWrite) throws IOException {
+        public int process(boolean shouldForceWrite,
+                           MutableLong lastForceWriteNanos) throws IOException {
             journalStats.getForceWriteQueueSize().dec();
             journalStats.getFwEnqueueTimeStats()
                 .registerSuccessfulEvent(MathUtils.elapsedNanos(enqueueTime), TimeUnit.NANOSECONDS);
@@ -380,10 +387,23 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
 
             try {
                 if (shouldForceWrite) {
-                    long startTime = MathUtils.nowInNano();
-                    this.logFile.forceWrite(false);
-                    journalStats.getJournalSyncStats()
-                        .registerSuccessfulEvent(MathUtils.elapsedNanos(startTime), TimeUnit.NANOSECONDS);
+                    // if entry can callback before forceWrite, trigger the callback first
+                    if (syncData
+                        || shouldClose
+                        || (MathUtils.elapsedMSec(lastForceWriteNanos.longValue()) >= journalForceWriteIntervalMs)) {
+                        for (int i = 0; i < forceWriteWaiters.size(); i++) {
+                            QueueEntry qe = forceWriteWaiters.get(i);
+                            if (qe != null && qe.ackBeforeSync) {
+                                forceWriteWaiters.set(i, null);
+                                cbThreadPool.execute(qe);
+                            }
+                        }
+                        long startTime = MathUtils.nowInNano();
+                        lastForceWriteNanos.setValue(startTime);
+                        this.logFile.forceWrite(false);
+                        journalStats.getJournalSyncStats()
+                            .registerSuccessfulEvent(MathUtils.elapsedNanos(startTime), TimeUnit.NANOSECONDS);
+                    }
                 }
                 lastLogMark.setCurLogMark(this.logId, this.lastFlushedPosition);
 
@@ -469,6 +489,7 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
         Thread threadToNotifyOnEx;
         // should we group force writes
         private final boolean enableGroupForceWrites;
+        private final MutableLong lastForceWriteNanos = new MutableLong();
         // make flush interval as a parameter
         public ForceWriteThread(Thread threadToNotifyOnEx, boolean enableGroupForceWrites) {
             super("ForceWriteThread");
@@ -489,6 +510,7 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
 
             boolean shouldForceWrite = true;
             int numReqInLastForceWrite = 0;
+            this.lastForceWriteNanos.setValue(MathUtils.nowInNano());
             while (running) {
                 ForceWriteRequest req = null;
                 try {
@@ -514,7 +536,7 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
                             }
                         }
                     }
-                    numReqInLastForceWrite += req.process(shouldForceWrite);
+                    numReqInLastForceWrite += req.process(shouldForceWrite, lastForceWriteNanos);
 
                     if (enableGroupForceWrites
                             // if its a marker we should switch back to flushing
@@ -610,6 +632,7 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
     private final boolean removePagesFromCache;
     private final int journalFormatVersionToWrite;
     private final int journalAlignmentSize;
+    private final int journalForceWriteIntervalMs;
 
     // Should data be fsynced on disk before triggering the callback
     private final boolean syncData;
@@ -647,13 +670,22 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
         super("BookieJournal-" + conf.getBookiePort());
         this.allocator = allocator;
 
+        final int numCallbackThreads = conf.getNumJournalCallbackThreads();
+
+        BlockingQueue<Runnable> cbQueue = new LinkedBlockingQueue<>();
         if (conf.isBusyWaitEnabled()) {
             // To achieve lower latency, use busy-wait blocking queue implementation
             queue = new BlockingMpscQueue<>(conf.getJournalQueueSize());
             forceWriteRequests = new BlockingMpscQueue<>(conf.getJournalQueueSize());
+            // if (numCallbackThreads > 0) {
+            //     cbQueue = new BlockingMpscQueue<>(conf.getJournalQueueSize() * 10);
+            // }
         } else {
             queue = new ArrayBlockingQueue<>(conf.getJournalQueueSize());
             forceWriteRequests = new ArrayBlockingQueue<>(conf.getJournalQueueSize());
+            // if (numCallbackThreads > 0) {
+            //     cbQueue = new ArrayBlockingQueue<>(conf.getJournalQueueSize() * 10);
+            // }
         }
 
         this.ledgerDirsManager = ledgerDirsManager;
@@ -670,9 +702,13 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
         this.bufferedEntriesThreshold = conf.getJournalBufferedEntriesThreshold();
         this.journalFormatVersionToWrite = conf.getJournalFormatVersionToWrite();
         this.journalAlignmentSize = conf.getJournalAlignmentSize();
-        if (conf.getNumJournalCallbackThreads() > 0) {
-            this.cbThreadPool = Executors.newFixedThreadPool(conf.getNumJournalCallbackThreads(),
-                                                         new DefaultThreadFactory("bookie-journal-callback"));
+        this.journalForceWriteIntervalMs = conf.getJournalForceWriteIntervalMs();
+        if (numCallbackThreads > 0) {
+            this.cbThreadPool = new ThreadPoolExecutor(
+                numCallbackThreads, numCallbackThreads,
+                0L, TimeUnit.MILLISECONDS,
+                cbQueue,
+                new DefaultThreadFactory("bookie-journal-callback"));
         } else {
             this.cbThreadPool = MoreExecutors.newDirectExecutorService();
         }
@@ -856,7 +892,7 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
     }
 
     @VisibleForTesting
-    void logAddEntry(long ledgerId, long entryId, ByteBuf entry, boolean ackBeforeSync, WriteCallback cb, Object ctx)
+    public void logAddEntry(long ledgerId, long entryId, ByteBuf entry, boolean ackBeforeSync, WriteCallback cb, Object ctx)
             throws InterruptedException {
         //Retain entry until it gets written to journal
         entry.retain();
@@ -884,6 +920,13 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
      */
     public int getJournalQueueLength() {
         return queue.size();
+    }
+
+    private void submitCallbacks(RecyclableArrayList<QueueEntry> entries) {
+        cbThreadPool.submit(() -> {
+            entries.forEach(QueueEntry::run);
+            entries.recycle();
+        });
     }
 
     /**
@@ -1004,7 +1047,7 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
                             // 2. If we have buffered more than the buffWriteThreshold or bufferedEntriesThreshold
                             shouldFlush = true;
                             journalStats.getFlushMaxOutstandingBytesCounter().inc();
-                        } else if (qe == null) {
+                        } else if (qe == null && flushWhenQueueEmpty) {
                             // We should get here only if we flushWhenQueueEmpty is true else we would wait
                             // for timeout that would put is past the maxWait threshold
                             // 3. If the queue is empty i.e. no benefit of grouping. This happens when we have one
@@ -1020,15 +1063,6 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
                             }
                             journalFlushWatcher.reset().start();
                             bc.flush();
-
-                            for (int i = 0; i < toFlush.size(); i++) {
-                                QueueEntry entry = toFlush.get(i);
-                                if (entry != null && (!syncData || entry.ackBeforeSync)) {
-                                    toFlush.set(i, null);
-                                    numEntriesToFlush--;
-                                    cbThreadPool.execute(entry);
-                                }
-                            }
 
                             lastFlushPosition = bc.position();
                             journalStats.getJournalFlushStats().registerSuccessfulEvent(
@@ -1050,26 +1084,12 @@ public class Journal extends BookieCriticalThread implements CheckpointSource {
                                 .registerSuccessfulValue(batchSize);
 
                             boolean shouldRolloverJournal = (lastFlushPosition > maxJournalSize);
-                            if (syncData) {
-                                // Trigger data sync to disk in the "Force-Write" thread.
-                                // Callback will be triggered after data is committed to disk
-                                forceWriteRequests.put(createForceWriteRequest(logFile, logId, lastFlushPosition,
-                                                                               toFlush, shouldRolloverJournal, false));
-                                toFlush = entryListRecycler.newInstance();
-                                numEntriesToFlush = 0;
-                            } else {
-                                // Data is already written on the file (though it might still be in the OS page-cache)
-                                lastLogMark.setCurLogMark(logId, lastFlushPosition);
-                                toFlush.clear();
-                                numEntriesToFlush = 0;
-                                if (shouldRolloverJournal) {
-                                    forceWriteRequests.put(
-                                            createForceWriteRequest(
-                                                    logFile, logId, lastFlushPosition,
-                                                    EMPTY_ARRAY_LIST, shouldRolloverJournal, false));
-                                }
-                            }
-
+                            // Trigger data sync to disk in the "Force-Write" thread.
+                            // Callback will be triggered after data is committed to disk
+                            forceWriteRequests.put(createForceWriteRequest(logFile, logId, lastFlushPosition,
+                                                                           toFlush, shouldRolloverJournal, false));
+                            toFlush = entryListRecycler.newInstance();
+                            numEntriesToFlush = 0;
                             batchSize = 0L;
                             // check whether journal file is over file limit
                             if (shouldRolloverJournal) {