@sijie 2019-08-23
I have used the journal perf tool to verify the behavior.

## without this change

### 1 dir

There is no performance difference (with sync and without sync) when there is only one journal on one disk.

- with sync

```
bin/bkperf journal append -j `pwd`/data1  -jt 1 -t 1 -gci 1 --sync
```

```
01:01:14,462 INFO  Throughput written : 593796.7  records/s ---    579.9 MB/s --- Latency: mean:  17.748 ms - med:  18.359 - 95pct:  23.145 - 99pct:  39.210 - 99.9pct: 186.220 - 99.99pct: 188.521 - Max: 188.823
01:01:24,485 INFO  Throughput written : 597023.1  records/s ---    583.0 MB/s --- Latency: mean:  18.437 ms - med:  17.752 - 95pct:  21.890 - 99pct:  29.045 - 99.9pct: 395.333 - 99.99pct: 397.813 - Max: 397.999
01:01:34,496 INFO  Throughput written : 620047.2  records/s ---    605.5 MB/s --- Latency: mean:  17.412 ms - med:  18.032 - 95pct:  24.668 - 99pct:  28.805 - 99.9pct:  38.970 - 99.99pct:  39.304 - Max:  39.491
01:01:44,513 INFO  Throughput written : 561966.1  records/s ---    548.8 MB/s --- Latency: mean:  19.213 ms - med:  18.208 - 95pct:  22.524 - 99pct:  28.528 - 99.9pct: 636.663 - 99.99pct: 638.127 - Max: 638.355
```

- without sync
```
bin/bkperf journal append -j `pwd`/data1  -jt 1 -t 1 -gci 1
```
```
01:00:13,268 INFO  Throughput written : 751407.2  records/s ---    733.8 MB/s --- Latency: mean:  13.542 ms - med:  10.793 - 95pct:  14.990 - 99pct:  70.896 - 99.9pct: 526.567 - 99.99pct: 527.255 - Max: 527.555
01:00:23,285 INFO  Throughput written : 703999.0  records/s ---    687.5 MB/s --- Latency: mean:  15.124 ms - med:  10.505 - 95pct:  15.918 - 99pct:  93.323 - 99.9pct: 678.683 - 99.99pct: 680.343 - Max: 680.511
01:00:33,304 INFO  Throughput written : 735972.5  records/s ---    718.7 MB/s --- Latency: mean:  13.734 ms - med:  10.623 - 95pct:  16.010 - 99pct:  18.623 - 99.9pct: 571.163 - 99.99pct: 571.983 - Max: 572.175
01:00:43,314 INFO  Throughput written : 714762.6  records/s ---    698.0 MB/s --- Latency: mean:  15.050 ms - med:  11.610 - 95pct:  16.562 - 99pct:  22.862 - 99.9pct: 570.799 - 99.99pct: 572.919 - Max: 573.007
```

### 8 dirs

The throughput decrease significantly when fsync is disabled.

- with sync
```
bin/bkperf journal append -j `pwd`/data1 -j `pwd`/data2 -j `pwd`/data3 -j `pwd`/data4 -j `pwd`/data5 -j `pwd`/data6 -j `pwd`/data7 -j `pwd`/data8 -jt 1 -t 1 -gci 1 --sync
```
```
01:03:09,006 INFO  Throughput written : 1221582.7  records/s ---   1193.0 MB/s --- Latency: mean:  44.899 ms - med:  44.498 - 95pct:  64.775 - 99pct: 158.941 - 99.9pct: 415.273 - 99.99pct: 420.547 - Max: 422.571
01:03:19,041 INFO  Throughput written : 1108485.6  records/s ---   1082.5 MB/s --- Latency: mean:  50.119 ms - med:  48.490 - 95pct:  92.226 - 99pct: 123.539 - 99.9pct: 470.601 - 99.99pct: 487.939 - Max: 492.791
01:03:29,065 INFO  Throughput written : 1146146.4  records/s ---   1119.3 MB/s --- Latency: mean:  46.478 ms - med:  43.410 - 95pct:  78.479 - 99pct: 384.447 - 99.9pct: 426.599 - 99.99pct: 430.683 - Max: 432.143
01:03:39,083 INFO  Throughput written : 1093618.1  records/s ---   1068.0 MB/s --- Latency: mean:  49.229 ms - med:  43.426 - 95pct:  76.283 - 99pct: 451.435 - 99.9pct: 610.411 - 99.99pct: 621.443 - Max: 629.307
01:03:49,104 INFO  Throughput written : 1144885.1  records/s ---   1118.1 MB/s --- Latency: mean:  45.297 ms - med:  44.397 - 95pct:  79.936 - 99pct: 198.625 - 99.9pct: 396.967 - 99.99pct: 406.649 - Max: 408.095
```

- without sync
```
bin/bkperf journal append -j `pwd`/data1 -j `pwd`/data2 -j `pwd`/data3 -j `pwd`/data4 -j `pwd`/data5 -j `pwd`/data6 -j `pwd`/data7 -j `pwd`/data8 -jt 1 -t 1 -gci 1
```
```
01:04:18,732 INFO  Throughput written : 360758.0  records/s ---    352.3 MB/s --- Latency: mean:  83.724 ms - med:  31.834 - 95pct: 353.531 - 99pct: 398.177 - 99.9pct: 507.453 - 99.99pct: 525.399 - Max: 526.015
01:04:28,772 INFO  Throughput written : 236682.2  records/s ---    231.1 MB/s --- Latency: mean: 196.678 ms - med: 189.764 - 95pct: 421.721 - 99pct: 473.305 - 99.9pct: 509.291 - 99.99pct: 529.231 - Max: 529.747
01:04:38,799 INFO  Throughput written : 302512.3  records/s ---    295.4 MB/s --- Latency: mean: 116.766 ms - med:  64.963 - 95pct: 374.395 - 99pct: 438.995 - 99.9pct: 484.735 - 99.99pct: 529.699 - Max: 663.183
01:04:48,819 INFO  Throughput written : 457502.7  records/s ---    446.8 MB/s --- Latency: mean:  11.671 ms - med:   0.922 - 95pct:  51.952 - 99pct: 247.162 - 99.9pct: 275.289 - 99.99pct: 299.569 - Max: 509.789
01:04:58,844 INFO  Throughput written : 238511.3  records/s ---    232.9 MB/s --- Latency: mean: 156.511 ms - med:  97.251 - 95pct: 445.239 - 99pct: 492.025 - 99.9pct: 526.227 - 99.99pct: 553.479 - Max: 553.567
01:05:08,863 INFO  Throughput written : 419129.7  records/s ---    409.3 MB/s --- Latency: mean:  37.548 ms - med:   1.079 - 95pct: 253.619 - 99pct: 405.653 - 99.9pct: 521.755 - 99.99pct: 571.427 - Max: 582.607
01:05:18,876 INFO  Throughput written : 404931.3  records/s ---    395.4 MB/s --- Latency: mean:  10.281 ms - med:   0.823 - 95pct:  30.702 - 99pct: 274.265 - 99.9pct: 498.939 - 99.99pct: 511.335 - Max: 511.491
```

---

With this change, we are able to get the same throughput on ssd even fsync is disabled.

```
01:09:44,040 INFO  Throughput written : 1091302.4  records/s ---   1065.7 MB/s --- Latency: mean:  48.060 ms - med:  43.743 - 95pct:  82.141 - 99pct: 124.934 - 99.9pct: 438.263 - 99.99pct: 525.255 - Max: 534.563
01:09:54,054 INFO  Throughput written : 1209835.9  records/s ---   1181.5 MB/s --- Latency: mean:  41.845 ms - med:  41.008 - 95pct:  74.531 - 99pct:  88.822 - 99.9pct: 102.985 - 99.99pct: 115.595 - Max: 118.077
01:10:04,072 INFO  Throughput written : 1134405.9  records/s ---   1107.8 MB/s --- Latency: mean:  46.664 ms - med:  45.284 - 95pct:  87.675 - 99pct: 107.904 - 99.9pct: 152.884 - 99.99pct: 186.559 - Max: 192.275
01:10:14,101 INFO  Throughput written : 823059.7  records/s ---    803.8 MB/s --- Latency: mean:  61.088 ms - med:  48.348 - 95pct: 142.495 - 99pct: 371.039 - 99.9pct: 577.211 - 99.99pct: 590.855 - Max: 591.651
01:10:24,131 INFO  Throughput written : 1027697.8  records/s ---   1003.6 MB/s --- Latency: mean:  53.062 ms - med:  45.302 - 95pct:  94.197 - 99pct: 502.447 - 99.9pct: 549.219 - 99.99pct: 556.727 - Max: 562.443
01:10:34,150 INFO  Throughput written : 1201110.2  records/s ---   1173.0 MB/s --- Latency: mean:  42.152 ms - med:  41.232 - 95pct:  71.689 - 99pct:  97.349 - 99.9pct: 352.749 - 99.99pct: 358.227 - Max: 359.385
```