@sijie 2019-08-23
run integration tests